Extract modules. refactor some code into methods. Reorganize code.
