+++ examples/tabbed-document-ui/src/documents/image.rs
@@ -1,0 +1,16 @@
+use std::path::PathBuf;
+
+pub struct ImageDocument {
+    path: PathBuf,
+    // TODO content: ImageContent(...)
+    coordinate: Option<(usize, usize)>
+}
+
+impl ImageDocument {
+    pub fn new(path: PathBuf) -> Self {
+        Self {
+            path,
+            coordinate: None,
+        }
+    }
+}
