+++ examples/tabbed-document-ui/src/documents/mod.rs
@@ -1,0 +1,10 @@
+use image::ImageDocument;
+use text::TextDocument;
+
+pub mod text;
+pub mod image;
+
+pub enum DocumentKind {
+    TextDocument(TextDocument),
+    ImageDocument(ImageDocument),
+}
