+++ examples/tabbed-document-ui/src/tabs/document.rs
@@ -1,0 +1,6 @@
+use slotmap::DefaultKey;
+
+#[derive(Clone)]
+pub struct DocumentTab {
+    pub document_key: DefaultKey,
+}
