+++ examples/tabbed-document-ui/src/tabs/home.rs
@@ -1,0 +1,3 @@
+#[derive(Clone)]
+pub struct HomeTab {
+}
